+++ src/lib/Scenes/Home/Components/ForYou/Components/__tests__/MarketingHeader-tests.tsx
@@ -1,0 +1,11 @@
+import React from "react"
+import "react-native"
+import * as renderer from "react-test-renderer"
+import { MarketingHeader } from "../MarketingHeader"
+
+describe("getSubjectArtist", () => {
+  it("looks correct when rendered", () => {
+    const tree = renderer.create(<MarketingHeader />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})