+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Master
 
 * Adding artwork badges to brick - matt
+* Add MarketingHeader to Home/WorksForYou - chris
 * Add @artsy/palette - yuki/orta/chris
 * Adds a Labs setting UI to the example app - orta
 * Updates Typescript version to 3.0.3